refactor : 로또게임(수동)
내용
* 로또 일급콜랙션에서 생성에 관련된 부분을 별도의 클래스로 분리 처리하고 테스트코드로 생성 테스트를 진행.
